tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animationevent_init_return.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - return value of initAnimationEvent</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animation-events-">
   <meta name="flags" content="dom">
   <meta name="assert" content="initAnimationEvent has return type void">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var log = document.querySelectorAll('p')[0].firstChild;
         var passed;
         try {
            var evt = document.createEvent('AnimationEvent');
            var returned = evt.initAnimationEvent('animationstart', true,
                                                   false, 'Test', 0);
            passed = (returned === undefined);
         } catch(e) {
            passed = false;
         }
         test(function(){assert_true(passed);}, testName); done();
         log.data = (passed ? 'PASS' : 'FAIL');
      };
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>Testing... (JavaScript required)</p>
</body>
</html>
